Minor bugfix
The pointers to the previous and next elements of an extracted item were
not previously cleared. This caused an error when displaying the
contents of a queue, since those pointers where used to follow up the
queued items.

Now the extracting functions clear the pointers to the previous and next
elements of the extracted elements before returning.